pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genesis"
version = "0.1.0"
description = "Multi-agent AI research experiment: cooperative identity formation, doctrine evolution, and ethical adaptation"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Alexander Fisher"}]

dependencies = [
    "pydantic>=2.0",
    "httpx>=0.25",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.2",
    "tiktoken>=0.5",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
include = ["controller*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
